+++ ui/ui.go
@@ -1794,7 +1794,7 @@ func packetNumberFromCurrentTableRow() (pcaptui.JumpPos, error) {
 
 func tableRowFromPacketNumber(savedPacket int) (int, error) {
 	// Map e.g. packet number #123 to the index in the PSML array - e.g. index 10 (order of psml load)
-	packetRowId, ok := Loader.PacketNumberMap[savedPacket]
+	packetRowId, ok := Loader.PacketRow(savedPacket)
 	if !ok {
 		return -1, fmt.Errorf("Error finding packet %v", savedPacket)
 	}
@@ -1983,7 +1983,7 @@ func vimKeysMainView(evk *tcell.EventKey, app gowid.IApp) bool {
 				savedPacket := lastJumpPos
 				if savedPacket != -1 {
 					// Map that packet number #123 to the index in the PSML array - e.g. index 10 (order of psml load)
-					if packetRowId, ok := Loader.PacketNumberMap[savedPacket]; !ok {
+					if packetRowId, ok := Loader.PacketRow(savedPacket); !ok {
 						OpenError(fmt.Sprintf("Error finding packet %v", savedPacket), app)
 					} else {
 						// This psml order is also the table RowId order. The table might be sorted though, so